palette multi select

+++ objects/o_dialog_color_selector/Create_0.gml
@@ -34,55 +34,57 @@ event_inherited();
 #region presets
 	preset_selecting = -1;
 	
-	sp_preset_w = ui(240 - 32 - 16);
-	sp_preset_size = ui(24);
-	click_block = false;
+	pal_padding    = ui(9);
+	sp_preset_w    = ui(240) - pal_padding * 2 - ui(8);
+	sp_preset_size = ui(20);
+	click_block    = false;
 	
-	sp_presets = new scrollPane(sp_preset_w, dialog_h - ui(62), function(_y, _m) {
-		var ww  = sp_preset_w - ui(40);
-		var hh  = ui(32);
-		var _gs = sp_preset_size;
-		var yy  = _y + ui(8);
-		var _hover = sHOVER && sp_presets.hover;
-		var _height, pre_amo;
+	sp_presets = new scrollPane(sp_preset_w, dialog_h - ui(48 + 8) - pal_padding, function(_y, _m) {
 		draw_clear_alpha(COLORS.panel_bg_clear, 0);
+		
+		var ww  = sp_presets.surface_w;
+		var _gs = sp_preset_size;
+		var hh  = ui(24);
+		var nh  = ui(20);
+		var pd  = ui(6);
+		var _ww = ww - pd * 2;
+		var _bh = nh + _gs + pd;
+		
+		var _hover = sHOVER && sp_presets.hover;
+		var _height, pre_amo, _palRes;
+		
+		var yy  = _y;
+		var col = max(1, floor(_ww / _gs)), row;
 		
 		for(var i = -1; i < array_length(PALETTES); i++) {
 			var pal = i == -1? {
-				name: "project",
-				palette: PROJECT.attributes.palette,
-				path: ""
+				name    : "project",
+				palette : PROJECT.attributes.palette,
+				path    : ""
 			} : PALETTES[i];
+			
 			pre_amo = array_length(pal.palette);
-			var col = floor(ww / _gs);
-			var row = ceil(pre_amo / col);
+			row     = ceil(pre_amo / col);
+			_height = preset_selecting == i? nh + row * _gs + pd : _bh;
 			
-			if(preset_selecting == i)
-				_height = ui(28) + row * _gs + ui(12);
-			else
-				_height = ui(56);
+			var isHover = _hover && point_in_rectangle(_m[0], _m[1], 0, yy, ww, yy + _height);
 			
-			var isHover = _hover && point_in_rectangle(_m[0], _m[1], ui(4), yy, ui(4) + sp_preset_w - ui(16), yy + _height);
-			
-			draw_sprite_stretched(THEME.ui_panel_bg, 3, ui(4), yy, sp_preset_w - ui(16), _height);
+			draw_sprite_stretched(THEME.ui_panel_bg, 3, 0, yy, ww, _height);
 			if(isHover) {
-				draw_sprite_stretched_ext(THEME.ui_panel_fg, 1, ui(4), yy, sp_preset_w - ui(16), _height, COLORS._main_accent, 1);
+				draw_sprite_stretched_ext(THEME.ui_panel_fg, 1, 0, yy, ww, _height, COLORS._main_accent, 1);
 				sp_presets.hover_content = true;
 			}
 			
-			var x0 = ui(16) + (i == -1) * ui(8 + 6);
 			var cc = i == preset_selecting? COLORS._main_accent : COLORS._main_text_sub;
 			draw_set_text(f_p2, fa_left, fa_top, cc);
-			draw_text(x0, yy + ui(8), pal.name);
-			if(i == -1) {
-				draw_set_color(cc);
-				draw_circle_prec(ui(16) + ui(4), yy + ui(16), ui(4), false);
-			}
+			draw_text_add(pd, yy + ui(2), pal.name);
+			
+			if(i == -1) { draw_set_color(cc); draw_circle_prec(ww - ui(10), yy + ui(10), ui(4), false); }
 			
 			if(preset_selecting == i) 
-				var _palRes = drawPaletteGrid(pal.palette, ui(16), yy + ui(28), ww, _gs, { color : selector.current_color, mx : _m[0], my : _m[1] });
+				_palRes = drawPaletteGrid(pal.palette, pd, yy + nh, _ww, _gs, { color : selector.current_color, mx : _m[0], my : _m[1] });
 			else
-				drawPalette(pal.palette, ui(16), yy + ui(28), ww, ui(20));
+				drawPalette(pal.palette, pd, yy + nh, _ww, _gs);
 			
 			if(!click_block && mouse_click(mb_left, interactable && sFOCUS)) {
 				if(preset_selecting == i && _hover && _palRes.hoverIndex > noone) {
@@ -95,20 +97,18 @@ event_inherited();
 				}
 			}
 			
-			if(isHover) {
-				if(i >= 0 && mouse_press(mb_right, interactable && sFOCUS)) {
-					hovering = pal;
-					
-					menuCall("palette_window_preset_menu",,, [
-						menuItem(__txtx("palette_editor_set_default", "Set as default"), function() { 
-							PROJECT.setPalette(array_clone(hovering.palette));
-						}),
-						menuItem(__txtx("palette_editor_delete", "Delete palette"), function() { 
-							file_delete(hovering.path); 
-							__initPalette();
-						}),
-					]);
-				}
+			if(isHover && i >= 0 && mouse_press(mb_right, interactable && sFOCUS)) {
+				hovering = pal;
+				
+				menuCall("palette_window_preset_menu",,, [
+					menuItem(__txtx("palette_editor_set_default", "Set as default"), function() { 
+						PROJECT.setPalette(array_clone(hovering.palette));
+					}),
+					menuItem(__txtx("palette_editor_delete", "Delete palette"), function() { 
+						file_delete(hovering.path); 
+						__initPalette();
+					}),
+				]);
 			}
 			
 			yy += _height + ui(4);
